api/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
filelock==3.14.0
flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.63.0
gunicorn==22.0.0
huggingface-hub==0.23.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
mido==1.3.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.20.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==23.2
pretty-midi==0.2.10
protobuf==3.20.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pyDeprecate==0.3.2
pyFluidSynth==1.3.3
python-dotenv==1.0.1
pytorch-lightning==1.6.5
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soundfile==0.12.1
sympy==1.12
tokenizers==0.12.1
torch==1.11.0
torchtext==0.12.0
tqdm==4.66.4
transformers==4.20.0
triton==2.3.0
typing-extensions==4.11.0
urllib3==2.2.1
werkzeug==3.0.2
yarl==1.9.4
zipp==3.18.1
